+++ CHANGELOG.md
@@ -152,8 +152,9 @@
   - Set `allowUnknownVariables: true` to [`no-invalid-properties`](https://github.com/eslint/css/blob/HEAD/docs/rules/no-invalid-properties.md)
   - Depending on the installed Tailwind version, enabled v3 or v4 Tailwind syntax support.
   - Added `allowedFeatures` option to conveniently set what CSS features will be ignored by [`use-baseline`](https://github.com/eslint/css/blob/HEAD/docs/rules/use-baseline.md) rule.
-- `typescript-eslint`: [8.32.0 → 8.46.3](https://github.com/typescript-eslint/typescript-eslint/compare/v8.32.0...v8.46.3)
+- `typescript-eslint`: [8.32.0 → 8.47.0](https://github.com/typescript-eslint/typescript-eslint/compare/v8.32.0...v8.47.0)
   - Set [`checkUnknown: true`](https://typescript-eslint.io/rules/no-base-to-string/#checkunknown) for [`no-base-to-string`](https://typescript-eslint.io/rules/no-base-to-string) rule.
+  - 🟢 (enabled) [`no-unused-private-class-members`](https://typescript-eslint.io/rules/no-unused-private-class-members)
 - `@eslint-react/eslint-plugin`: [1.49.0 → 2.3.1](https://github.com/Rel1cx/eslint-react/compare/v1.49.0...v2.3.1)
   - [**BREAKING**] Some rules have been renamed and removed. Please refer to [v2 migration guide](https://github.com/Rel1cx/eslint-react/releases/tag/v2.0.0) for more details.
   - [**BREAKING**] Debug rules have been moved to a separate [`eslint-plugin-react-debug`](https://npmjs.com/eslint-plugin-react-debug) package.